Use correct conventions & move commands to separate classes.
